Break up a huge file
