+++ src/module/dataforged/index.ts
@@ -1,6 +1,4 @@
-import { ItemDataConstructorData } from '@league-of-foundry-developers/foundry-vtt-types/src/foundry/common/data/data.mjs/itemData'
-import { IronswornActor } from '../actor/actor'
-import { get, isArray, isObject, max, set } from 'lodash'
+import { isArray, isObject } from 'lodash'
 import {
   starforged,
   ironsworn,
@@ -10,12 +8,12 @@ import {
   Starforged,
   Ironsworn,
 } from 'dataforged'
-import { marked } from 'marked'
 import { IronswornItem } from '../item/item'
 import shajs from 'sha.js'
 import { cachedDocumentsForPack } from '../features/pack-cache'
-import { RollTableDataConstructorData } from '@league-of-foundry-developers/foundry-vtt-types/src/foundry/common/data/data.mjs/rollTableData.js'
-import { TableResultDataConstructorData } from '@league-of-foundry-developers/foundry-vtt-types/src/foundry/common/data/data.mjs/tableResultData.js'
+
+export * from './import'
+export * from './rendering'
 
 // For some reason, rollupJs mangles this
 const SFMoveCategories = ((starforged as any).default as Starforged)[
@@ -28,22 +26,6 @@ const ISOracleCategories = ((ironsworn as any).default as Ironsworn)[
   'Oracle Categories'
 ]
 const SFAssetTypes = ((starforged as any).default as Starforged)['Asset Types']
-
-function getLegacyRank(numericRank) {
-  switch (numericRank) {
-    case 1:
-      return 'troublesome'
-    case 2:
-      return 'dangerous'
-    case 3:
-      return 'formidable'
-    case 4:
-      return 'extreme'
-    case 5:
-      return 'epic'
-  }
-  return 'epic'
-}
 
 export function cleanDollars(obj): any {
   if (isArray(obj)) {
@@ -155,336 +137,4 @@ export function findOracleWithIntermediateNodes(
     walkCategory(cat)
   }
   return ret
-}
-
-const COMPENDIUM_KEY_MAP = {
-  'Ironsworn/Moves': 'ironswornmoves',
-  'Ironsworn/Oracles': 'ironswornoracles',
-  'Starforged/Moves': 'starforgedmoves',
-  'Starforged/Oracles': 'starforgedoracles',
-  'Starforged/Encounters': 'starforgedencounters',
-}
-const MARKDOWN_LINK_RE = /\[(.*?)\]\((.*?)\)/g
-const DESCRIPTOR_FOCUS_RE = /\[Descriptor \+ Focus\]\(.*?\)/
-const ACTION_THEME_RE = /\[Action \+ Theme\]\(.*?\)/
-
-function idIsOracleLink(dfid: string): boolean {
-  return /^(Starforged|Ironsworn)\/Oracle/.test(dfid)
-}
-
-export function renderLinksInStr(text: string): string {
-  // Strip "Black Medium Right-Pointing Triangle" characters
-  text = text.replace('\u23f5', '')
-
-  // Strip brackets from e.g. factions/name template
-  text = text.replace(/\[(\[.*?\))\]/g, '$1')
-
-  // Catch "Descriptor+Focus" or "Action+Theme" and replace with two links
-  text = text.replace(
-    DESCRIPTOR_FOCUS_RE,
-    '[Descriptor](Starforged/Oracles/Core/Descriptor) + [Focus](Starforged/Oracles/Core/Focus)'
-  )
-  text = text.replace(
-    ACTION_THEME_RE,
-    '[Action](Starforged/Oracles/Core/Action) + [Theme](Starforged/Oracles/Core/Theme)'
-  )
-
-  return text.replace(MARKDOWN_LINK_RE, (match, text, url) => {
-    const parts = url.split('/')
-    const kind = `${parts[0]}/${parts[1]}`
-    const compendiumKey = COMPENDIUM_KEY_MAP[kind]
-    if (!compendiumKey) return match
-    if (idIsOracleLink(url)) {
-      return `<a class="entity-link oracle-category-link" data-dfid="${url}"><i class="fa fa-caret-right"></i> ${text}</a>`
-    }
-    return `@Compendium[foundry-ironsworn.${compendiumKey}.${hash(
-      url
-    )}]{${text}}`
-  })
-}
-
-function renderMarkdown(md: string, markedFn = marked.parse) {
-  return markedFn(renderLinksInStr(md))
-}
-
-export function renderLinksInMove(move: IMove) {
-  const textProperties = [
-    'Text',
-    'Trigger.Text',
-    'Outcomes.Strong Hit.Text',
-    'Outcomes.Strong Hit.With a Match.Text',
-    'Outcomes.Weak Hit.Text',
-    'Outcomes.Miss.Text',
-    'Outcomes.Miss.With a Match.Text',
-  ]
-  for (const prop of textProperties) {
-    const text = get(move, prop)
-    if (!text) continue
-    set(move, prop, renderLinksInStr(text))
-  }
-}
-
-const PACKS = [
-  'foundry-ironsworn.starforgedassets',
-  'foundry-ironsworn.starforgedencounters',
-  'foundry-ironsworn.starforgedmoves',
-  'foundry-ironsworn.starforgedoracles',
-  'foundry-ironsworn.foeactorssf',
-  'foundry-ironsworn.ironswornoracles',
-]
-
-/**
- * Converts JSON from dataforged resources into foundry packs. Requires packs to
- * already exist, but will empty them prior to repopulating. In a perfect world
- * we would retain dataforged's $id as Foundry's _id, but Foundry validates that
- * those fields meet with its UID expectations (/[a-zA-Z0-9]{16}/)
- */
-export async function importFromDataforged() {
-  // Empty out the packs
-  for (const key of PACKS) {
-    const pack = game.packs.get(key)
-    if (!pack) continue
-
-    // Unlock all the packs
-    await pack.configure({ locked: false })
-
-    // @ts-ignore IdQuery type is a little bogus
-    const idsToDelete = pack.index.map((x) => x._id)
-    await Item.deleteDocuments(idsToDelete, { pack: key })
-  }
-
-  await processSFMoves()
-  await processSFAssets()
-  await processSFOracles()
-  await processSFEncounters()
-  await processSFFoes()
-
-  await processISOracles()
-
-  // Lock the packs again
-  for (const key of PACKS) {
-    await game.packs.get(key)?.configure({ locked: true })
-  }
-}
-
-async function processSFMoves() {
-  const movesToCreate = [] as (ItemDataConstructorData &
-    Record<string, unknown>)[]
-  for (const category of SFMoveCategories) {
-    for (const move of category.Moves) {
-      renderLinksInMove(move)
-      const cleanMove = cleanDollars(move)
-      movesToCreate.push({
-        _id: hashLookup(cleanMove['dfid']),
-        type: 'sfmove',
-        name: cleanMove['Name'],
-        img: 'icons/dice/d10black.svg',
-        data: cleanMove,
-      })
-    }
-  }
-  await Item.createDocuments(movesToCreate, {
-    pack: 'foundry-ironsworn.starforgedmoves',
-    keepId: true,
-  })
-}
-
-async function processSFAssets() {
-  const assetsToCreate = [] as (ItemDataConstructorData &
-    Record<string, unknown>)[]
-  for (const assetType of SFAssetTypes) {
-    for (const asset of assetType.Assets) {
-      assetsToCreate.push({
-        type: 'asset',
-        _id: hashLookup(asset.$id),
-        name: `${assetType.Name} / ${asset.Name}`,
-        data: {
-          description: renderMarkdown(assetType.Description),
-          fields:
-            asset.Inputs?.map((input) => ({
-              name: input.Name,
-              value: '',
-            })) || [],
-          abilities: (asset.Abilities ?? []).map((ability) => {
-            const ret = {
-              enabled: ability.Enabled || false,
-              description: renderMarkdown(ability.Text),
-            } as any
-
-            for (const input of ability.Inputs ?? []) {
-              if (input['Input Type'] === 'Clock') {
-                const ic = input
-                ret.hasClock = true
-                ret.clockMax = ic.Segments
-                ret.clockTicks = ic.Filled
-              }
-              // TODO: other input types
-            }
-
-            return ret
-          }),
-          track: {
-            enabled: !!asset['Condition Meter'],
-            name: asset['Condition Meter']?.Name,
-            current: asset['Condition Meter']?.Value,
-            max: asset['Condition Meter']?.Max,
-          },
-          exclusiveOptions: [], // TODO:
-        },
-      })
-    }
-  }
-  await Item.createDocuments(assetsToCreate, {
-    pack: 'foundry-ironsworn.starforgedassets',
-    keepId: true,
-  })
-}
-
-/**
- * ORACLES
- */
-async function processOracle(
-  oracle: IOracle,
-  output: RollTableDataConstructorData[]
-) {
-  // Oracles JSON is a tree we wish to iterate through depth first adding
-  // parents prior to their children, and children in order
-  if (oracle.Table) {
-    const description = marked.parseInline(
-      renderLinksInStr(oracle.Description ?? '')
-    )
-    const maxRoll = max(oracle.Table.map((x) => x.Ceiling || 0)) //oracle.Table && maxBy(oracle.Table, (x) => x.Ceiling)?.Ceiling
-    output.push({
-      _id: hashLookup(oracle.$id),
-      flags: {
-        dfId: oracle.$id,
-        category: oracle.Category,
-      },
-      name: oracle.Name,
-      img: 'icons/dice/d10black.svg',
-      description,
-      formula: `d${maxRoll}`,
-      replacement: true,
-      displayRoll: true,
-      /* folder: // would require using an additional module */
-      results: oracle.Table?.map((tableRow) => {
-        let text: string
-        if (tableRow.Result && tableRow.Summary) {
-          text = `${tableRow.Result} (${tableRow.Summary})`
-        } else text = tableRow.Result ?? ''
-        return {
-          _id: hashLookup(tableRow.$id ?? ''),
-          range: [tableRow.Floor, tableRow.Ceiling],
-          text: tableRow.Result && renderLinksInStr(text),
-        } as TableResultDataConstructorData
-      }).filter((x) => x.range[0] !== null),
-    })
-  }
-
-  for (const child of oracle.Oracles ?? []) await processOracle(child, output)
-}
-async function processOracleCategory(
-  cat: IOracleCategory,
-  output: RollTableDataConstructorData[]
-) {
-  for (const oracle of cat.Oracles ?? []) await processOracle(oracle, output)
-  for (const child of cat.Categories ?? [])
-    await processOracleCategory(child, output)
-}
-
-async function processSFOracles() {
-  const oraclesToCreate: RollTableDataConstructorData[] = []
-
-  for (const category of SFOracleCategories) {
-    await processOracleCategory(category, oraclesToCreate)
-  }
-  await RollTable.createDocuments(oraclesToCreate, {
-    pack: 'foundry-ironsworn.starforgedoracles',
-    keepId: true,
-  })
-}
-
-async function processISOracles() {
-  const oraclesToCreate: RollTableDataConstructorData[] = []
-
-  for (const category of ISOracleCategories) {
-    await processOracleCategory(category, oraclesToCreate)
-  }
-  await RollTable.createDocuments(oraclesToCreate, {
-    pack: 'foundry-ironsworn.ironswornoracles',
-    keepId: true,
-  })
-}
-
-async function processSFEncounters() {
-  const encountersToCreate = [] as (ItemDataConstructorData &
-    Record<string, unknown>)[]
-  for (const encounter of starforged.Encounters) {
-    const description = await renderTemplate(
-      'systems/foundry-ironsworn/templates/item/sf-foe.hbs',
-      {
-        ...encounter,
-        variantLinks: encounter.Variants.map((x) =>
-          renderLinksInStr(`[${x.Name}](${x.$id})`)
-        ),
-      }
-    )
-
-    encountersToCreate.push({
-      _id: hashLookup(encounter['$id']),
-      type: 'progress',
-      name: encounter['Name'],
-      data: {
-        description,
-        rank: getLegacyRank(encounter['Rank']),
-      },
-    })
-
-    for (const variant of encounter['Variants']) {
-      const variantDescription = await renderTemplate(
-        'systems/foundry-ironsworn/templates/item/sf-foe.hbs',
-        {
-          ...encounter,
-          ...variant,
-          Category: variant['Nature'] || encounter['Nature'],
-          CategoryDescription: variant['Summary'] || encounter['Summary'],
-        }
-      )
-
-      encountersToCreate.push({
-        _id: hashLookup(variant['$id']),
-        type: 'progress',
-        name: variant['Name'],
-        data: {
-          description: variantDescription,
-          rank: getLegacyRank(
-            'Rank' in variant ? variant['Rank'] : encounter['Rank']
-          ),
-        },
-      })
-    }
-  }
-  await Item.createDocuments(encountersToCreate, {
-    pack: 'foundry-ironsworn.starforgedencounters',
-    keepId: true,
-  })
-}
-
-async function processSFFoes() {
-  const foesPack = game.packs.get('foundry-ironsworn.starforgedencounters')
-  const foeItems =
-    (await foesPack?.getDocuments()) as StoredDocument<IronswornItem>[]
-  for (const foeItem of foeItems ?? []) {
-    const actor = await IronswornActor.create(
-      {
-        name: foeItem.name ?? 'wups',
-        img: foeItem.data.img,
-        type: 'foe',
-      },
-      { pack: 'foundry-ironsworn.foeactorssf' }
-    )
-    await actor?.createEmbeddedDocuments('Item', [
-      foeItem.data as unknown as Record<string, unknown>,
-    ])
-  }
 }
